// src/app/data/enums/modeltype.enum.ts
export enum Modeltype {
    Vendor = "Vendor",
    Client = "Client",
    Community = "Community",
    OrganisationTranslation = "OrganisationTranslation",
    SuccessStory = "SuccessStory",
    SuccessStoryTranslation = "SuccessStoryTranslation",
    Product = "Product",
    ProductTranslation = "ProductTranslation",
    ProductCategory = "ProductCategory",
    ProductCategoryTranslation = "ProductCategoryTranslation",
    VendorProductServices = "VendorProductServices",
    Industry = "Industry",
    IndustryTranslation = "IndustryTranslation",
    Language = "Language",
    Address = "Address",
    CommunityMember = "CommunityMember",
    NewsFeed = "NewsFeed",
    TopNews = "TopNews"
}